pyproject.toml
[project]
name = "automatic-template-generation"
version = "0.1.0"
description = "Extract design elements from flat images and return them in Polotno-compatible JSON format"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
    "python-multipart>=0.0.12",
    "jinja2>=3.1.0",
    "openai>=1.55.0",
    "google-generativeai>=0.8.0",
    "pillow>=11.0.0",
    "numpy>=2.0.0",
    "scipy>=1.14.0",
    "scikit-image>=0.24.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.28.0",
    "google-genai>=1.52.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
